+++ examples/settings.py
@@ -1,6 +1,12 @@
 import logging
+import os
 import sys
 
+from microagent import configure
+
+
+cur_dir = os.path.dirname(os.path.realpath(__file__))
+configure('file://' + os.path.join(cur_dir, 'signals.json'))
 logging.basicConfig(format=(
     '%(levelname)-8s [pid#%(process)d] %(asctime)s %(name)s '
     '%(filename)s:%(lineno)d %(message)s'
@@ -11,7 +17,6 @@ BUS = {
     'redis': {
         'backend': 'microagent.tools.redis.RedisSignalBus',
         'dsn': 'redis://localhost/7',
-        'prefix': 'PREF',
     },
 }