+++ tests/conftest.py
@@ -1,5 +1,7 @@
 import pytest
 
+import microagent.bus as bus_module
+
 from microagent import Queue, Signal
 
 
@@ -7,3 +9,4 @@ from microagent import Queue, Signal
 async def flush_signals_and_queues() -> None:
     Queue._queues = {}
     Signal._signals = {}
+    bus_module._DEFAULT_PREFIX = 'PUBSUB'